# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Beehive Manager API
  description: API for managing agents and tasks in the Beehive system
  version: 1.0.0
servers:
  - url: https://{server}:8080
    variables:
      server:
        default: localhost
        description: Server hostname
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
  schemas:
...
    TaskOutput:
      type: object
      properties:
        logs:
          type: string
        error:
          type: string
    Task:
      type: object
      required:
        - agent_id
        - type
        - parameters
      properties:
        id:
          type: string
        agent_id:
          type: string
        type:
          type: string
          enum: [command_shell, file_operation, ui_automation, browser_automation]
        parameters:
          type: object
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled, timeout]
        output:
          $ref: '#/components/schemas/TaskOutput'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        timeout:
          type: integer
        started_at:
          type: string
          format: date-time
    TaskCreationResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [queued]
        timestamp:
          type: string
          format: date-time
    TaskCancelResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [cancelled]
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /api/task/create:
    post:
      summary: Create a new task
      security:
        - ApiKeyAuth: []
        - SignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task
              properties:
                task:
                  $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreationResponse'
        '400':
          description: Invalid request
        '401':
          description: Authentication failed
  /api/task/status/{task_id}:
    get:
      summary: Get task status
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
        - SignatureAuth: []
      responses:
        '200':
          description: Task status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
  /api/task/cancel/{task_id}:
    post:
      summary: Cancel a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
        - SignatureAuth: []
      responses:
        '200':
          description: Task cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCancelResponse'
        '404':
          description: Task not found
...